requirements.txt
Flask==3.0.0
requests==2.31.0
python-telegram-bot==20.7
gunicorn==21.2.0